+++ packages/atomic/src/components/search/atomic-tab-manager/atomic-tab-manager.spec.ts
@@ -41,12 +41,14 @@ describe('atomic-tab-manager', () => {
     const {element} = await renderInAtomicSearchInterface<AtomicTabManager>({
       template: html`
         <atomic-tab-manager ?clear-filters-on-tab-change=${clearFiltersOnTabChange}>
-          ${slottedContent ??
-          html`
-            <atomic-tab label="All" name="all"></atomic-tab>
-            <atomic-tab label="Documentation" name="documentation"></atomic-tab>
-            <atomic-tab label="Articles" name="articles"></atomic-tab>
-          `}
+          ${
+            slottedContent ??
+            html`
+              <atomic-tab label="All" name="all"></atomic-tab>
+              <atomic-tab label="Documentation" name="documentation"></atomic-tab>
+              <atomic-tab label="Articles" name="articles"></atomic-tab>
+            `
+          }
         </atomic-tab-manager>
       `,
       selector: 'atomic-tab-manager',