+++ packages/quantic/force-app/main/default/lwc/quanticFoldedResultList/e2e/fixture.ts
@@ -36,6 +36,7 @@ export const testSearch =
       await use(new SearchObject(page, searchRequestRegex));
     },
     resultList: async ({page, options, configuration, search}, use) => {
+      await search.mockSearchWithBaseResponse();
       await page.goto(pageUrl);
       configuration.configure(options);
       await search.waitForSearchResponse();
@@ -57,11 +58,14 @@ export const testInsight =
       {page, options, search, configuration, insightSetup},
       use
     ) => {
+      await search.mockSearchWithBaseResponse();
       await page.goto(pageUrl);
       configuration.configure({...options, useCase: useCaseEnum.insight});
       await insightSetup.waitForInsightInterfaceInitialization();
-      await search.performSearch();
-      await search.waitForSearchResponse();
+      await Promise.all([
+        search.waitForSearchResponse(),
+        search.performSearch(),
+      ]);
       await search.waitForSearchResultsVisible();
       await use(new FoldedResultListObject(page));
     },