+++ packages/quantic/force-app/main/default/lwc/quanticSmartSnippet/e2e/fixture.ts
@@ -43,8 +43,8 @@ export const testSearch =
     ) => {
       const smartSnippetObject = new SmartSnippetObject(page);
 
-      await page.goto(pageUrl);
       await search.mockSearchWithSmartSnippetResponse(data);
+      await page.goto(pageUrl);
 
       configuration.configure(options);
       await search.waitForSearchResponse();
@@ -77,13 +77,15 @@ export const testInsight =
     ) => {
       const smartSnippetObject = new SmartSnippetObject(page);
 
-      await page.goto(pageUrl);
       await search.mockSearchWithSmartSnippetResponse(data);
+      await page.goto(pageUrl);
 
       configuration.configure({...options, useCase: useCaseEnum.insight});
       await insightSetup.waitForInsightInterfaceInitialization();
-      await search.performSearch();
-      await search.waitForSearchResponse();
+      await Promise.all([
+        search.waitForSearchResponse(),
+        search.performSearch(),
+      ]);
       await use(smartSnippetObject);
     },
   });